:recycle: Resolved content not matching error
by nesting suspense on dynamic component
